+++ src/main/java/hotspot/user/policy/controller/response/DayBlockedTimeResponse.java
@@ -1,0 +1,12 @@
+package hotspot.user.policy.controller.response;
+
+import java.time.DayOfWeek;
+import java.util.List;
+
+import lombok.Builder;
+
+@Builder
+public record DayBlockedTimeResponse(
+        DayOfWeek day,               // 요일
+        List<BlockedRangeResponse> ranges // 차단 시간대
+) {}
